+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Implement nunito font and remove unused font files
 - When user pause the video show display ad after 10 sec
 - Add more menu
+- Add report screen in inactive state
 
 ## v2.0.23.15
 - Reduce progress bar, and fix it for IE11